requirements.txt
altgraph==0.17.3
argparse-addons==0.12.0
astroid==2.15.6
attrs==23.1.0
bitstruct==8.17.0
cantools==38.0.2
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
crccheck==1.3.0
cycler==0.11.0
dill==0.3.7
diskcache==5.6.1
exceptiongroup==1.1.1
fonttools==4.39.4
importlib-resources==5.12.0
iniconfig==2.0.0
isort==5.12.0
jsonschema==4.17.3
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
matplotlib==3.7.1
mccabe==0.7.0
numpy==1.24.3
packaging==23.1
pandas==2.0.2
pefile==2023.2.7
Pillow==9.5.0
platformdirs==3.10.0
pluggy==1.2.0
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
pylint==2.17.5
pyparsing==3.0.9
PyQt6==6.4.2
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
PyQt6-sip==13.5.1
pyqt6-tools==6.4.2.3.3
pyrsistent==0.19.3
pytest==7.4.0
pytest-mock==3.11.1
pytest-qt==4.2.0
python-can==4.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.0
qt6-applications==6.4.3.2.3
qt6-tools==6.4.3.1.3
scipy==1.10.1
six==1.16.0
textparser==0.24.0
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.6.3
tzdata==2023.3
wrapt==1.15.0
zipp==3.15.0
